Add a bootstrap script that downloads the project
The bootstrap script is designed to work with downloading the project
both from GitHub and from a locally stored version

The flight-direct server's IP is stored as an environment variable. This
way it is easier to change it's value at a latter date. It is also a
easy way to distinguish between a local network and github build