Add some additional CLI helper functionality
There is a somewhat standardized way we run our commands. The syntax
follows a predictable pattern which is now constructed by the
`default_syntax` method.

Also the name of the command corresponds to the Class name within the
Commands module. Thus this can be implicitly determined as well.

See comments in the code for more details.