Change the entry point into the program
A new `flight_direct.rb` file has been created and shall act as the
base file for the project. It is responsible for setting the load path
and setting up `Bundler`. It also requires the base set of gems that
will always be required (plus the `commands`).

The original `bin/flight` now requires the above file and runs the CLI.
This way the requiring of the project can be separated from the running
of the CLI.

The `CLI` is no longer responsible for requiring anything accept for
`Commander`. Thus it should be possible to skip it entirely if need be