Add a base `profile.d` script that runs the project
The `etc` directory is now installed in with the project. The post
install script already sources `etc/profile.d` directory.

All that needed to be defined now is a bash function that runs the
`bin/flight` ruby script.

The PATH has been updated to use the `embedded/bin` that was created by
the omnibus installer