Define a development build with readline support
`readline` is licensed under GPL and thus we might have issues using it
in the future.

Currently it is only used by some of the development gems, as such a
`flight-direct-dev` build has been created. This is for development
purposes only and will not form part of the final product.

This does however compile ruby with `readline` support which is off by
default.