Use the root directory env variable to set ruby paths
This should allow for the root to be changed in development. This means
an installed `rpm` can run the development code with ease

In short, the `FLIGHT_DIRECT_ROOT` is used to load the base
`flight_direct.rb` file.

However, before ruby's LOAD_PATH is set, it checks for a special
`FLIGHT_DIRECT_RUBY_SOURCE` environment variable. This path is used in
development mode to load a different set of source code.

It is ignored outside of `dev` mode and defaults to the regular root