Remove the hard coded ruby path
During the `ruby` compilation, it installs the standard suite of ruby
commands to:
`embedded/bin/{irb, gem, ....}`

However it also hard codes the shebangs in these files to:
`#!/opt/flight-direct/embedded/bin/ruby`

This means ruby is not portable as these files will not exist unless
installed to the `install_dir`.

The fix is to `sed` the hard coded shebangs and point them to:
`/usr/bin/env ruby`. This way they will always look at the system path