Update the ruby environment variables
The `GEM_HOME` needs to be a single directory so it acts as a fall back
location. `GEM_PATH` then needs to add it to the front of it's path so
it takes priority.

Finally, `flight-direct` should share its gems which have been bundle
installed. Thus the `BUNDLE_PATH` points to `vendor/cache`