Set the clusterware library path
Clusterware is use to be at the root and thus it can work out its
function path. However, for the purposes of `flight`, it is installed to
`opt/clusterware/functions`. (This is the only part of `clusterware`
that is required by `flight`).

As such the `cw_LIBRARY` path needs to point to this new location