Replace the old Commander cli with Thor
Commander is complicated and has does things automatically making it
hard to expand on.

Thor, on the other hand, hasn't been built on OptionParser nor ARGV. It
also natively supports sub-commanding.

For this reason the old CLI has been removed and replaced. It currently
is in its infancy, so watch this space