Alias `run!` to be the same as `start`
This alias has been done intentionally. The `start` method is an
internal `Thor` method that runs the command. However it preprocessing
is required, then `super` would have to be used.

Instead the legacy `run!` method shall be used instead. Currently it is
an alias, however it can be expanded in the future.