Set the GEM environment variables during installation
The GEM environment variables make it tricky to determine the install
location of the gems.

So they are now being step explicitly to be within the `install_dir`.
`flight` does not use the shared directory as it uses a standalone
install. This removes the runtime requirement on `Bundler` but also
means it generates a special `setup.rb` script that takes its place.