Automatically pass on `--help` in help command
The `help` command has been overridden to parse on `--help` flags to the
other commands/ actions. This way, each action is responsible for its
own help.

Otherwise there would be two levels of help, one provided by Thor and
other by the action.

This is done by using the inbuilt `Thor#invoke` method

NOTE: This is a special case when the `options` hash is not empty. As
the commands are defined with `*args` as inputs, they can wrap any
number of inputs. It also causes `Thor` to preserve the option flags in
the `*args` array. This means `options` is typically empty.

However when `invoke` is called in the `help` method, it does add the
"help" flag to the internal `options` list and does not pass it in with
the `args`. This means a special case is required to append it to the
array.