Add versioning information into flight direct
Instead of always building off the local directory, versioning has
been added to the project.

From now on, the default version is what ever has been set in the new
`version.rb` file. For a regular production build, this version needs to
correspond with a git tag.

The development build overrides the version however and will continue
to build from the local source.