Make a production-vm Vagrantfile
A dedicated Vagrantfile has been created to deploy `flight`. This will
prevent it getting ran by accident as you will have to run `vagrant up`
within the directory.

The provisioning script will run all the commands required to deploy and
publish to S3. This will generate a consistent deployment environment in
which to build