Put the Bash in'ERB in'Bash in'Ruby in'Vagrant
It is not a good idea to let different commits of `flight` to build
arbitrary tags. Also the git directory should be clean as well.

BUT Vagrant....

The check shouldn't be done on all the vagrant commands (e.g.
vagrant destroy). Instead it should only be done when the provisioning
script is ran.

So ERB?

ERB is a easy way to achieve this, as it can conditional bake in the
error when appropriate. The conditions will be evaluated on the host
machine when the Vagrantfile is being parsed. However the error will
only be raised if the provisioning file is ran.

Scary GIT-ERB-BASH!!!!

You may notice there is embedded shell-git commands inside ERB tags that
renders a bash script. Because the ERB is being renderd, those git
commands run on the host machine up front.

This means the decision to error has been made well before it actually
happens. So remember kids, look both ways when crossing the street.