Create a development bin for handy scripts
There are a few tasks that I use regularly, a prime example is
installing rvm and the development gems.

As this commands are only used in development, they have been added to a
`vagrant_bin` directory. A setup script inside this directory will add
it to the PATH. This is done automatically on the master `vm`