Mount the omnibus cache into the master vm
Currently omnibus caches in the vm's home directory. The issue with
doing this is it makes the cache ephemeral.

However this means everything needs to be compiled from source each time
a new vm is created. Having a persistent cache will speed this process
up by reducing the boiler plate compiling of git, OpenSSL etc