Export the flight "aliases" into the env
The `alces` alias to `flight` is used in some of the scripts associated
with gridware. If theses where to be run in a subshell created by
flight, they would fail.

Instead, the aliases are dynamically defined as a function and exported
into the environment. This means the aliases will always be available.

At the same time, the "bash" action should no longer source the base.sh
file. Instead it shouldn't need to source anything again.