Allow the command details to use #: as the delim
The existing delimiter ':' is not valid ruby and thus needs to be
commented out. For this reason, '#:' can optionally be used as the
delimiter as well.

Otherwise the same rules apply

NOTE: There has been a bug fix where one of the regex matches was
missing the beginning of line check