Define files in `libexec/thor` as Loki commands
Any files placed in the `libexec/thor` directory will be parsed as
Thor code. This is done with the new Loki parser, and thus no boiler
plate code is required.

Subcommands have explicitly NOT be used. This is because subcommands
need to be required up front each time a command is ran. This will slow
down `flight` significantly if a large plugin is included.

Instead, the thor plugin is only parsed if that command has been
executed. This means the plugin files are only required when needed