Create a Loki::ThorExt module to help build Thor
There are various CLI features which are common to a lot of apps. As
such a `Loki::ThorExt` module has been created. Currently this module
only has a method to reset the help command.

NOTE: It is safe to rely on `super` over Thor public methods. These
methods define the command names and thus can not be altered without
creating a breaking change.