Parse all lines that start with '#:' for cmd info
This block is a useful way to parse information about the command to the
parent without requiring/running the file.

However, the 'law of least surprise' would expect it to read the entire
declaration and stop on the first regular line. In practice, a blank
line should always follow the declaration

The '#' optional is optional and is missing extensively in the legacy code.
This should be updated.

Shebangs are allowed with the declaration