Formalise the current code as a `jit_parse` method
Essentially current code parses the subcommands in a `JIT` manner. It is
only after it has selected which command to run, does it parse the code.

Doing this requires both class and instance methods, and thus the module
now needs to be included not extended.

Also, this moves all the parsing code within Loki, allowing it to be
reused