Preserve existing config values
Before exporting the new values, they are merged over the old config
values. This will prevent old values from being lost even if they aren't
exported into the environment correctly.

It also means it isn't possible to change the config by altering the
environment, however it is possible to add a new cofig. This is not a
supported feature