Render the cron file into each time directory
The use of anacron complicates the setup of the `flight/etc/cron.*`
directories. Anacron has a minimum time resolution of 1 day and is thus
only used for the longer time periods. It has the advantage that it
works after a reboot.

This does mean the shorter time periods (e.g. an hour) need to be
handled by cron itself. To use both features requires at least two
system config files.

As such it is easier to render all the files at once