Implicitly add the packages libraries to load path
The `FlightSyncer` forge package has its own libraries which will be
stored within `flight-direct/lib`. This means they are implicitly on
the load path ONCE INSTALLED.

However this leaves the development libraries off the path. To fix this,
the development libraries path is inferred from the cli file. It is then
added to the LOAD_PATH if it exists