Implement a Loki Command to provide error handling
Thor only allows for simple error handling of number of arguments. Loki
will extend this functionality to allow for `--trace` to be implemented.

Loki will thus need to run the provided block in `rescue`'s. To achieve
this, the provided block is turned into a private `_run_loki` method.
Then the public method can be used to implement the error handling.

NOTE: `define_method` only creates public methods that can then be made
private. This causes Thor to issue a warning about the public method.
The warning has thus been suppressed with the `no_commands` block